pyproject.toml
[project]
name = "pythonanywhere-dispenser-apis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-ninja>=1.4.0",
    "dotenv>=0.9.9",
    "pdf2image>=1.16.3",
    "Pillow>=10.0.0",
]
